+++ src/core/hle/kernel/process.h
@@ -12,7 +12,7 @@
 #include <boost/container/static_vector.hpp>
 #include "common/bit_field.h"
 #include "common/common_types.h"
-#include "core/hle/kernel/kernel.h"
+#include "core/hle/kernel/object.h"
 #include "core/hle/kernel/vm_manager.h"
 
 namespace Kernel {
