+++ src/core/hle/kernel/semaphore.h
@@ -7,7 +7,7 @@
 #include <string>
 #include <queue>
 #include "common/common_types.h"
-#include "core/hle/kernel/kernel.h"
+#include "core/hle/kernel/object.h"
 #include "core/hle/kernel/wait_object.h"
 #include "core/hle/result.h"
