+++ src/core/hle/kernel/server_port.h
@@ -8,7 +8,7 @@
 #include <string>
 #include <tuple>
 #include "common/common_types.h"
-#include "core/hle/kernel/kernel.h"
+#include "core/hle/kernel/object.h"
 #include "core/hle/kernel/wait_object.h"
 
 namespace Kernel {
