+++ src/core/hle/kernel/server_session.h
@@ -8,7 +8,7 @@
 #include <string>
 #include "common/assert.h"
 #include "common/common_types.h"
-#include "core/hle/kernel/kernel.h"
+#include "core/hle/kernel/object.h"
 #include "core/hle/kernel/wait_object.h"
 #include "core/hle/result.h"
 #include "core/memory.h"
